moved QueueFactory to more appropriate place
