+++ pyjob/__init__.py
@@ -26,21 +26,7 @@ __version__ = version.__version__
 
 from pyjob.cexec import cexec
 from pyjob.pool import Pool
+from pyjob.queue import QueueFactory
 
 
-def QueueFactory(platform, *args, **kwargs):
-    """Accessibility function for any :obj:`Queue <pyjob.queue.Queue>`
-    
-    Parameters
-    ----------
-    platform : str
-       The platform to create the queue on
 
-    """
-    platform = platform.lower()
-    if platform == 'local':
-        from pyjob.local import LocalJobServer
-        return LocalJobServer(*args, **kwargs)
-    elif platform == 'sge':
-        from pyjob.sge import SunGridEngine
-        return SunGridEngine(*args, **kwargs)
